[fix] only render indicator when available
