+++ packages/next/src/next-devtools/dev-overlay/dev-overlay.tsx
@@ -5,13 +5,11 @@ import { ComponentStyles } from './styles/component-styles'
 import { CssReset } from './styles/css-reset'
 import { Colors } from './styles/colors'
 import { ErrorOverlay } from './components/errors/error-overlay/error-overlay'
-import { DevToolsIndicator } from './components/errors/dev-tools-indicator/dev-tools-indicator'
 import { RenderError } from './container/runtime-error/render-error'
 import { DarkTheme } from './styles/dark-theme'
-import { useDevToolsScale } from './components/errors/dev-tools-indicator/dev-tools-info/preferences'
 import type { HydrationErrorState } from '../shared/hydration-error'
 import type { ReadyRuntimeError } from './utils/get-error-by-type'
-import { DevToolsIndicatorNew } from './components/devtools-indicator/devtools-indicator'
+import { DevToolsIndicator } from './components/devtools-indicator/devtools-indicator'
 import { PanelRouter } from './menu/panel-router'
 import { PanelRouterContext, type PanelStateKind } from './menu/context'
 import type { OverlayState, OverlayDispatch } from './shared'
@@ -32,9 +30,7 @@ export function DevOverlay({
   dispatch: OverlayDispatch
   getSquashedHydrationErrorDetails: (error: Error) => HydrationErrorState | null
 }) {
-  const [scale, setScale] = useDevToolsScale()
   const [panel, setPanel] = useState<null | PanelStateKind>(null)
-  const isBuildError = state.buildError !== null
   const [selectedIndex, setSelectedIndex] = useState(-1)
 
   const triggerRef = useRef<HTMLButtonElement>(null)
@@ -74,32 +70,11 @@ export function DevOverlay({
                         errorCount={totalErrorCount}
                       />
                       <PanelRouter />
-                      <DevToolsIndicatorNew />
+                      <DevToolsIndicator />
                     </PanelRouterContext>
                   </RenderErrorContext>
                 </>
-              ) : (
-                <>
-                  <DevToolsIndicator
-                    scale={scale}
-                    setScale={setScale}
-                    state={state}
-                    dispatch={dispatch}
-                    errorCount={totalErrorCount}
-                    isBuildError={isBuildError}
-                  />
-
-                  <ErrorOverlay
-                    state={state}
-                    dispatch={dispatch}
-                    getSquashedHydrationErrorDetails={
-                      getSquashedHydrationErrorDetails
-                    }
-                    runtimeErrors={runtimeErrors}
-                    errorCount={totalErrorCount}
-                  />
-                </>
-              )}
+              ) : null}
             </>
           )
         }}
